refactor: 보유 와인에 빈티지 적용
