+++ src/main/java/com/drinkeg/drinkeg/domain/myWine/dto/request/MyWineUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.drinkeg.drinkeg.domain.myWine.controller.request;
+package com.drinkeg.drinkeg.domain.myWine.dto.request;
 
 import lombok.*;
 
@@ -8,11 +8,13 @@ import java.time.LocalDate;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class MyWineUpdateRequest {
 
+    private Integer vintageYear;
     private LocalDate purchaseDate;
     private Integer purchasePrice;
 
     @Builder
-    public MyWineUpdateRequest(LocalDate purchaseDate, Integer purchasePrice) {
+    public MyWineUpdateRequest(Integer vintageYear, LocalDate purchaseDate, Integer purchasePrice) {
+        this.vintageYear = vintageYear;
         this.purchaseDate = purchaseDate;
         this.purchasePrice = purchasePrice;
     }
